# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reading Log API",
    "version": "1.0.0",
    "description": "API documentation for your Reading Log project"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/book": {
      "get": {
        "summary": "Get all books belonging to the logged-in user",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all books for the logged-in user"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Add a new book for the logged-in user",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "authorFirstName": {
                    "type": "string"
                  },
                  "authorLastName": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "publishedDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "pages": {
                    "type": "integer"
                  },
                  "readStatus": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book added successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/book/{id}": {
      "get": {
        "summary": "Get a single book by ID (only if owned by the logged-in user)",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book found"
          },
          "400": {
            "description": "Invalid book ID"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Book not found or unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update a book by ID (only if owned by the logged-in user)",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "400": {
            "description": "Invalid input or book ID"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Book not found or unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a book by ID (only if owned by the logged-in user)",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          },
          "400": {
            "description": "Invalid book ID"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Book not found or unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {},
  "tags": [
    {
      "name": "Books",
      "description": "API for managing books in your reading log"
    }
  ]
}